server/pyproject.toml
[project]
name = "tak-manager-server"
version = "3.1.2"
description = "Backend for TAK Manager"
authors = [
    {name = "JShadowNull",email = "95039572+JShadowNull@users.noreply.github.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<1.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "pexpect (>=4.9.0,<5.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "lxml (>=5.3.1,<6.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
